fix(ephys): use int32 for sample_count and UTC-aware index in synced_df
- Change OnixImuChunk.sample_count from bare `int` to `int32` for
  consistency with sister stream tables (EnvironmentState, MessageLog,
  LightEvents) in acquisition.py
- Replace per-element _harp_to_naive loop in synced_df with vectorized
  io_api.to_datetime(harp_seconds), which returns a UTC-aware
  DatetimeIndex as required by the spec
- Tighten test_synced_df_returns_harp_indexed_dataframe to assert
  df.index.tz is not None (UTC-aware per spec)
